* commit '8477a24660f705d7e36350294b274b5a8058286d': (702 commits)
  [EDELIVERY-5507] fix Sonar error
  [EDELIVERY-5576] logging no more necessary
  [EDELIVERY-5576] use DomainTaskExecutor task for getQueueSize
  [EDELIVERY-5576] added debug
  [EDELIVERY-5571] - Default value for chunking set to false
  [EDELIVERY-5571] - Default value for chunking set to false
  [EDELIVERY-5566] address pull request comments
  [EDELIVERY-5566] address pull request comments
  [EDELIVERY-5566] address pull request comments
  [EDELIVERY-5566] added a property for show only DLQ
  [EDELIVERY-5566] added back normal Gauge if refreshPeriod == 0
  [EDELIVERY-5566] fixed multitenancy + JMS count only for DLQ
  [EDELIVERY-5551] fixed JUnit
  [EDELIVERY-5557] added info into changelog.txt file
  [EDELIVERY-5557] added TODO
  [EDELIVERY-5415] - Changed version to 4.1.2
  [EDELIVERY-5415] - Changed version to 4.1.2
  [EDELIVERY-5415] - Removed settings.xml
  [EDELIVERY-5415] - Updated upgrade-info.txt
  [EDELIVERY-5415] - Updated changelog.txt
  ...

+++ Domibus-MSH/src/main/java/eu/domibus/common/dao/ConfigurationRawDAO.java
@@ -25,7 +25,7 @@ public class ConfigurationRawDAO extends BasicDao<ConfigurationRaw> {
         super(ConfigurationRaw.class);
     }
 
-    public ConfigurationRaw getConfigurationRaw(int id) {
+    public ConfigurationRaw getConfigurationRaw(long id) {
         final TypedQuery<ConfigurationRaw> query = this.em.createNamedQuery("ConfigurationRaw.getById", ConfigurationRaw.class);
         query.setParameter("CONF_ID", id);
         return query.getSingleResult();
